+++ src/main/java/com/asg/finance/dto/BankPaymentItemDetailRequest.java
@@ -9,12 +9,12 @@ public class BankPaymentItemDetailRequest {
     private Long detRowId;
     private Long stockPoid;
     private Long stockUnitPoid;
-    private Long poQty;
-    private Long dnQty;
-    private Long qtyReceived;
-    private Long price;
-    private Long discount;
-    private Long total;
+    private Double poQty;
+    private Double dnQty;
+    private Double qtyReceived;
+    private Double price;
+    private Double discount;
+    private Double total;
     private String remarks;
     private String refDocId;
     private Long refDocPoid;
